+++ strongtalk/src/cpp/main/vm/compiler/BitVector.hpp
@@ -15,32 +15,32 @@ class LongRegisterMask;
 // General bit vectors; Space is allocated for maxLength bits, but only the portion up to length bits is used.
 // SimpleBitVector: simple version that fits into a word (for speed)
 
-typedef void (*intDoFn)( std::int32_t i );
+typedef void (*intDoFn)( std::size_t i );
 
-// a pseudo-class -- std::int32_t(this) actually holds the bits, so there's no allocation
+// a pseudo-class -- std::size_t(this) actually holds the bits, so there's no allocation
 class SimpleBitVector : public ValueObject {
 
-    std::int32_t _bits;
+    std::size_t _bits;
 
 public:
-    SimpleBitVector( std::int32_t b = 0 ) :
+    SimpleBitVector( std::size_t b = 0 ) :
         _bits{ b } {
     }
 
 
-    SimpleBitVector allocate( std::int32_t l ) {
+    SimpleBitVector allocate( std::size_t l ) {
         st_assert( l < BITS_PER_WORD, "need longer bit vector" );
         return SimpleBitVector( addNthBit( _bits, l ) );
     }
 
 
-    SimpleBitVector deallocate( std::int32_t l ) {
+    SimpleBitVector deallocate( std::size_t l ) {
         st_assert( l < BITS_PER_WORD, "need longer bit vector" );
         return SimpleBitVector( subNthBit( _bits, l ) );
     }
 
 
-    bool isAllocated( std::int32_t l ) {
+    bool isAllocated( std::size_t l ) {
         st_assert( l < BITS_PER_WORD, "need longer bit vector" );
         return isBitSet( _bits, l );
     }
@@ -56,55 +56,55 @@ public:
 class BitVector : public PrintableResourceObject {
 
 protected:
-    std::int32_t _maxLength; // max # bits
+    std::size_t _maxLength; // max # bits
     std::size_t length;     // number of bits, not words
-    std::int32_t *_bits;     // array containing the bits
+    std::size_t *_bits;     // array containing the bits
 
-    std::int32_t indexFromNumber( std::int32_t i ) {
+    std::size_t indexFromNumber( std::size_t i ) {
         return i >> LOG_2_BITS_PER_WORD;
     }
 
 
-    std::int32_t offsetFromNumber( std::int32_t i ) {
+    std::size_t offsetFromNumber( std::size_t i ) {
         return lowerBits( i, LOG_2_BITS_PER_WORD );
     }
 
 
-    bool getBitInWord( std::int32_t i, std::int32_t o ) {
+    bool getBitInWord( std::size_t i, std::size_t o ) {
         return isBitSet( _bits[ i ], o );
     }
 
 
-    void setBitInWord( std::int32_t i, std::int32_t o ) {
+    void setBitInWord( std::size_t i, std::size_t o ) {
         setNthBit( _bits[ i ], o );
     }
 
 
-    void clearBitInWord( std::int32_t i, std::int32_t o ) {
+    void clearBitInWord( std::size_t i, std::size_t o ) {
         clearNthBit( _bits[ i ], o );
     }
 
 
-    std::int32_t bitsLength( std::int32_t l ) {
+    std::size_t bitsLength( std::size_t l ) {
         return indexFromNumber( l - 1 ) + 1;
     }
 
 
-    std::int32_t *createBitString( std::int32_t l ) {
-        std::int32_t blen = bitsLength( l );
-        std::int32_t *bs  = new_resource_array<std::int32_t>( blen );
+    std::size_t *createBitString( std::size_t l ) {
+        std::size_t blen = bitsLength( l );
+        std::size_t *bs  = new_resource_array<std::size_t>( blen );
         set_words( bs, blen, 0 );
         return bs;
     }
 
 
-    std::int32_t *copyBitString( std::int32_t len ) {
+    std::size_t *copyBitString( std::size_t len ) {
 
         st_assert( len >= _maxLength, "can't shorten" );
 
-        std::int32_t blen    = bitsLength( len );
-        std::int32_t *bs     = new_resource_array<std::int32_t>( blen );
-        std::int32_t blength = bitsLength( _maxLength );
+        std::size_t blen    = bitsLength( len );
+        std::size_t *bs     = new_resource_array<std::size_t>( blen );
+        std::size_t blength = bitsLength( _maxLength );
         copy_words( _bits, bs, blength );
         if ( blength < blen ) {
             set_words( bs + blength, blen - blength, 0 );
@@ -115,7 +115,7 @@ protected:
 
 
 public:
-    BitVector( std::int32_t l ) :
+    BitVector( std::size_t l ) :
         _maxLength{ l },
         length{ l },
         _bits{ createBitString( l ) } {
@@ -132,7 +132,7 @@ public:
 
 
 protected:
-    BitVector( std::int32_t l, std::int32_t ml, std::int32_t *bs ) :
+    BitVector( std::size_t l, std::size_t ml, std::size_t *bs ) :
         _maxLength{ ml },
         length{ l },
         _bits{ bs } {
@@ -140,7 +140,7 @@ protected:
 
 
 public:
-    BitVector *copy( std::int32_t len ) {
+    BitVector *copy( std::size_t len ) {
         return new BitVector( length, len, copyBitString( len ) );
     }
 
@@ -160,16 +160,16 @@ public:
     }
 
 
-    void addFromTo( std::int32_t first, std::int32_t last );    // set bits [first..last]
+    void addFromTo( std::size_t first, std::size_t last );    // set bits [first..last]
 
-    void remove( std::int32_t i ) {
+    void remove( std::size_t i ) {
         //st_assert( this, "shouldn't be a null pointer" );
         st_assert( i < length, "not in range" );
         clearBitInWord( indexFromNumber( i ), offsetFromNumber( i ) );
     }
 
 
-    void removeFromTo( std::int32_t first, std::int32_t last );    // clear bits [first..last]
+    void removeFromTo( std::size_t first, std::size_t last );    // clear bits [first..last]
 
     // union/intersect return true if receiver has changed
     bool unionWith( BitVector *other );          // this |= other
@@ -178,7 +178,7 @@ public:
 
     void doForAllOnes( intDoFn f );  // call f for all 1 bits
 
-    void setLength( std::int32_t l ) {
+    void setLength( std::size_t l ) {
         st_assert( l < _maxLength, "too big" );
         length = l;
     }
@@ -195,5 +195,5 @@ public:
 
     friend class LongRegisterMask;
 
-    friend std::int32_t findFirstUnused( LongRegisterMask **strings, std::int32_t len, std::int32_t start );
+    friend std::size_t findFirstUnused( LongRegisterMask **strings, std::size_t len, std::size_t start );
 };
